rename lstm_eval to lstm_eval_internal
